requirements.txt
discord.py>=1.7.0
aiohttp>=3.7.0
redis>=4.2.0
